chore: refactor workspace types
