+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/index.ts
@@ -1,4 +1,4 @@
 export * from './functions';
 export * from './SideSheets';
 export * from './WorkspaceConfig';
-
+export * from './options';
