+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/dataSource.ts
@@ -1,0 +1,6 @@
+export type DataSource<T extends Record<PropertyKey, unknown>> = {
+    /** Function that returns the api call promise */
+    responseAsync: (signal?: AbortSignal) => Promise<Response>;
+    /** Function that parses the response to correct format, defaults to just parsing the raw response */
+    responseParser?: (Response: Response) => Promise<T[]>;
+};
