+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/dataViewerProps.ts
@@ -1,0 +1,5 @@
+import { ViewOptions } from './viewOptions';
+
+export type DataViewerProps<T extends Record<PropertyKey, unknown>> = ViewOptions<T> & {
+    data: T;
+};
