+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/helpPageOptions.ts
@@ -1,0 +1,3 @@
+export type HelpPageOptions = {
+    Component: () => JSX.Element;
+};
