+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/powerBIOptions.ts
@@ -1,0 +1,7 @@
+import { Filter, PBIOptions } from '@equinor/lighthouse-powerbi';
+
+export type PowerBiOptions = {
+    reportURI: string;
+    filter?: Filter[];
+    options?: PBIOptions;
+};
