+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/prefetchQueriesOptions.ts
@@ -1,0 +1,6 @@
+import { FetchQueryOptions, QueryFunction } from 'react-query';
+export interface PrefetchQueriesOptions {
+    queryKey: string[];
+    queryFn: QueryFunction<unknown, string[]>;
+    options?: FetchQueryOptions<unknown, unknown, unknown, string[]> | undefined;
+}
