+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/searchOptions.ts
@@ -1,0 +1,5 @@
+export type SearchOption<T> = {
+    name: string;
+    /** Takes in an item and returns the search value */
+    valueFormatter: (item: T) => string;
+};
