+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/statusFunc.ts
@@ -1,0 +1,2 @@
+import { StatusItem } from '@equinor/lighthouse-status-bar';
+export type StatusFunc<T extends Record<PropertyKey, unknown>> = (data: T[]) => StatusItem[];
