+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/timelineOptions.ts
@@ -1,0 +1,1 @@
+export interface TimeLineOptions {}
