+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/validator.ts
@@ -1,0 +1,1 @@
+export type Validator<T> = (data: unknown[]) => T[];
