+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/viewOptions.ts
@@ -1,0 +1,11 @@
+export type ViewOptions<T extends Record<PropertyKey, unknown>> = {
+    objectIdentifierKey: keyof T;
+    title?: {
+        key: keyof T;
+        label: string;
+    };
+    description?: {
+        key: keyof T;
+        label: string;
+    };
+};
