+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/workflowEditorOptions.ts
@@ -1,0 +1,3 @@
+export interface WorkflowEditorOptions {
+    endpoint: string;
+}
