+++ src/Core/WorkSpace/src/WorkSpaceApi/Types/options/workspaceTab.ts
@@ -1,0 +1,1 @@
+export type WorkspaceTab = 'tree' | 'table' | 'garden' | 'analytics' | 'gantt' | 'editor' | 'help';
