+++ src/Core/WorkSpace/src/index.ts
@@ -1,7 +1,6 @@
 export * from './WorkSpace';
 export * from './WorkSpaceApi';
 export * from './WorkSpaceApi/useWorkSpace';
-export type { TableOptions, TreeOptions } from './WorkSpaceApi/workspaceState';
 export * from './WorkSpaceApi/WorkSpaceTypes';
 
 export * from './Components';
