+++ src/apps/Punch_Dev/utility/config/analyticsConfig.ts
@@ -1,4 +1,4 @@
-import { PowerBiOptions } from '../../../../Core/WorkSpace/src/WorkSpaceApi/workspaceState';
+import { PowerBiOptions } from '@equinor/WorkSpace';
 
 export const analyticsConfig: PowerBiOptions = {
     reportURI: 'pp-punch-analytics',
