+++ src/apps/swcr/utilities/tableSetup.tsx
@@ -1,5 +1,5 @@
 import { Column } from '@equinor/Table';
-import { TableOptions } from '../../../Core/WorkSpace/src/WorkSpaceApi/workspaceState';
+import { TableOptions } from '@equinor/WorkSpace';
 import { SwcrPackage } from '../models/SwcrPackage';
 import { getNextSignatureRoleKeys, getNextToSignKeys, getTypeKeys } from './packages';
 export const tableConfig: TableOptions<SwcrPackage> = {
@@ -171,7 +171,7 @@ export const columns: Column<SwcrPackage>[] = [
             const count = row.subRows.reduce((acc, curr) => {
                 acc =
                     curr.original.status !== 'Closed' &&
-                        curr.original.status !== 'Closed - Rejected'
+                    curr.original.status !== 'Closed - Rejected'
                         ? acc + 1
                         : acc;
                 return acc;
@@ -194,7 +194,7 @@ export const columns: Column<SwcrPackage>[] = [
             const count = row.subRows.reduce((acc, curr) => {
                 acc =
                     curr.original.status === 'Closed' ||
-                        curr.original.status === 'Closed - Rejected'
+                    curr.original.status === 'Closed - Rejected'
                         ? acc + 1
                         : acc;
                 return acc;
